storage: split disk related types
Remove cache dependency on whole storage lib
