+++ src/v/cloud_storage/tests/BUILD
@@ -52,6 +52,7 @@ redpanda_test_cc_library(
         "//src/v/model/tests:random",
         "//src/v/redpanda/tests:fixture",
         "//src/v/storage",
+        "//src/v/storage:disk",
         "//src/v/storage:record_batch_utils",
         "//src/v/storage/tests:disk_log_builder",
         "//src/v/test_utils:fixture",
@@ -267,6 +268,7 @@ redpanda_cc_btest(
         "//src/v/cloud_storage:cache",
         "//src/v/config",
         "//src/v/random:generators",
+        "//src/v/storage:disk",
         "//src/v/test_utils:iostream",
         "//src/v/test_utils:scoped_config",
         "//src/v/test_utils:seastar_boost",
@@ -664,6 +666,7 @@ redpanda_cc_gtest(
         "//src/v/random:generators",
         "//src/v/redpanda/tests:fixture",
         "//src/v/storage",
+        "//src/v/storage:disk",
         "//src/v/test_utils:gtest",
         "//src/v/test_utils:scoped_config",
         "@googletest//:gtest",
