+++ src/v/cluster/BUILD
@@ -742,6 +742,7 @@ redpanda_cc_library(
         "//src/v/ssx:sleep_abortable",
         "//src/v/ssx:when_all",
         "//src/v/storage",
+        "//src/v/storage:disk",
         "//src/v/storage:parser_utils",
         "//src/v/storage:record_batch_builder",
         "//src/v/storage:record_batch_utils",
