+++ src/v/cluster/node/local_monitor.cc
@@ -18,8 +18,8 @@
 #include "config/configuration.h"
 #include "config/node_config.h"
 #include "ssx/future-util.h"
+#include "storage/disk.h"
 #include "storage/node.h"
-#include "storage/types.h"
 #include "utils/human.h"
 #include "version/version.h"
