+++ src/v/storage/types.cc
@@ -45,18 +45,6 @@ std::optional<kafka::offset> stm_manager::lowest_pinned_data_offset() const {
     return result;
 }
 
-std::ostream& operator<<(std::ostream& o, const disk& d) {
-    fmt::print(
-      o,
-      "{{path: {}, free: {}, total: {}, alert: {}, fsid: {}}}",
-      d.path,
-      human::bytes(d.free),
-      human::bytes(d.total),
-      d.alert,
-      d.fsid);
-    return o;
-}
-
 std::ostream& operator<<(std::ostream& o, const log_reader_config& cfg) {
     o << "{start_offset:" << cfg.start_offset
       << ", max_offset:" << cfg.max_offset << ", min_bytes:" << cfg.min_bytes
